converter/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "e2neutrino"
version = "0.1.0"
description = "Convert Enigma2 settings into Neutrino (Zapit) profiles."
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Neutrino Settings Automation", email = "team@example.com" }
]
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
  "requests>=2.31",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "ruff>=0.5.0",
  "mypy>=1.11",
  "build>=1.2.1",
  "types-PyYAML>=6.0.0.20240904",
]

[project.scripts]
e2neutrino = "e2neutrino.__main__:main"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
